Доп. прическа, доработка логики обновления

+++ src/Services/ProductService/ProductService.Application/ProductManagementService.cs
@@ -4,18 +4,11 @@ using SharedLibrary.ProductService.Dto;
 
 namespace ProductService.Application
 {
-    public class ProductManagementService : IProductManagementService
+    public class ProductManagementService(IProductRepository repository, IFileStorage fileStorage, IMapper mapper) : IProductManagementService
     {
-        private readonly IProductRepository _repository;
-        private readonly IFileStorage _fileStorage;
-        private readonly IMapper _mapper;
-
-        public ProductManagementService(IProductRepository repository, IFileStorage fileStorage, IMapper mapper)
-        {
-            _repository = repository;
-            _fileStorage = fileStorage;
-            _mapper = mapper;
-        }
+        private readonly IProductRepository _repository = repository;
+        private readonly IFileStorage _fileStorage = fileStorage;
+        private readonly IMapper _mapper = mapper;
 
         public Task<List<Product>> GetAllAsync(CancellationToken ct = default) => _repository.GetAllAsync(ct);
 
@@ -34,20 +27,30 @@ namespace ProductService.Application
             return product.Id;
         }
 
-        public async Task<string> UpdateAsync(ProductUpdateDto dto, CancellationToken ct = default)
+        public async Task UpdateAsync(string id, ProductUpdateDto dto, CancellationToken ct = default)
         {
-            var product = _mapper.Map<Product>(dto);
+            var product = await _repository.GetByIdAsync(id, ct);
+            if (product is null)
+                return;
+
+            //Создаем сущность нового продукта, отдаем ей айдишник и присваиваем старую (возможную) картинку
+            var newProduct = _mapper.Map<Product>(dto);
+            newProduct.Id = id;
+            newProduct.ImageUrl = product.ImageUrl;
+
+            //Если в dto передана картинка
             if (dto.Image != null)
             {
+                //В БД уже есть УРЛ старой картинки
                 if (!string.IsNullOrEmpty(product.ImageUrl))
-                {
+                    //Чистим старую
                     await _fileStorage.DeleteFileAsync(product.ImageUrl);
-                    product.ImageUrl = string.Empty;
-                }
-                product.ImageUrl = await _fileStorage.UploadFileAsync(dto.Image, ct);
+
+                //Добавляем новую
+                newProduct.ImageUrl = await _fileStorage.UploadFileAsync(dto.Image, ct);
             }
-            await _repository.UpdateAsync(product, ct);
-            return product.Id;
+
+            await _repository.UpdateAsync(newProduct, ct);
         }
 
         public async Task DeleteAsync(string id, CancellationToken ct = default)
